+++ src/uploads/validators.py
@@ -5,25 +5,25 @@ def validate_text(field):
     ext = os.path.splitext(field.name)[-1]
     valid_extensions = ['.pdf']
     if not ext.lower() in valid_extensions:
-        raise ValidationError('Unsupported file extension. Must be pdf.')
+        raise ValidationError('Unsupported file extension. Must be .pdf.')
 
 def validate_video(field):
     ext = os.path.splitext(field.name)[-1]
-    valid_extensions = ['.mp4', '.mpeg4']
+    valid_extensions = ['.mp4', '.mpeg4', '.mov']
     if not ext.lower() in valid_extensions:
-        raise ValidationError('Unsupported file extension. Must be mp4.')
+        raise ValidationError(f'Unsupported file extension. Must be one of {", ".join(valid_extensions)}.')
 
 def validate_audio(field):
     ext = os.path.splitext(field.name)[-1]
     valid_extensions = ['.mp3', '.mpeg3', '.m4a', '.wav']
     if not ext.lower() in valid_extensions:
-        raise ValidationError('Unsupported file extension. Must be one of %s.' % ', '.join(valid_extensions))
+        raise ValidationError(f'Unsupported file extension. Must be one of {", ".join(valid_extensions)}.')
 
 def validate_captions(field):
     ext = os.path.splitext(field.name)[-1]
     valid_extensions = ['.vtt', '.srt']
     if not ext.lower() in valid_extensions:
-        raise ValidationError('Unsupported file extension. Must be vtt or srt.')
+        raise ValidationError(f'Unsupported file extension. Must be .vtt or .srt.')
 
 def validate_barcode(field):
     try: